+++ projects/gmpm/mpm/PPW.cu
@@ -187,7 +187,8 @@ struct ZSParticleParticleWrangler : INode {
           (unsigned short)unitBytes,
           (unsigned short)tileSize,
           (unsigned short)targetParPtr->numChannels(),
-          (unsigned short)targetParPtr->getChannelOffset(name) + dimid,
+          (unsigned short)targetParPtr->getChannelOffset(name) +
+              (unsigned short)dimid,
           (unsigned short)isNeighborProperty};
     }
     auto daccessors = haccessors.clone({zs::memsrc_e::device, 0});
